+++ channel-simulation/java-channel-simulation/src/main/java/modulator/Modulator.java
@@ -3,13 +3,13 @@ package modulator;
 import util.Buffer;
 
 public abstract class Modulator {
-    protected final double carrierFrequency, modulationFrequency, carrierAmplitude;
+    protected final float carrierFrequency, modulationFrequency, carrierAmplitude;
     public static final String[] parameters = {"carrier frequency", "modulation frequency", "carrier amplitude"};
     public final Buffer buffer = new Buffer();
 
-    public abstract double[] calculate(byte[] data, double timeStep);
+    public abstract float[] calculate(byte[] data, float timeStep);
 
-    public Modulator(double carrierFrequency, double modulationFrequency, double carrierAmplitude) {
+    public Modulator(float carrierFrequency, float modulationFrequency, float carrierAmplitude) {
         this.carrierFrequency = carrierFrequency;
         this.modulationFrequency = modulationFrequency;
         this.carrierAmplitude = carrierAmplitude;
