+++ channel-simulation/java-channel-simulation/src/main/java/modulator/ModulatorFactory.java
@@ -5,10 +5,10 @@ import java.util.ArrayList;
 public class ModulatorFactory {
 
     public static ASKModulator createASKModulator(ArrayList<String> parameters) {
-        double carrierFrequency = Double.parseDouble(parameters.get(0));
-        double modulationFrequency = Double.parseDouble(parameters.get(1));
-        double carrierAmplitude = Double.parseDouble(parameters.get(2));
-        double depth = Double.parseDouble(parameters.get(3));
+        float carrierFrequency = Float.parseFloat(parameters.get(0));
+        float modulationFrequency = Float.parseFloat(parameters.get(1));
+        float carrierAmplitude = Float.parseFloat(parameters.get(2));
+        float depth = Float.parseFloat(parameters.get(3));
 
         return new ASKModulator(carrierFrequency, modulationFrequency, carrierAmplitude, depth);
     }
